+++ packages/block-library/src/post-date/edit.js
@@ -10,13 +10,12 @@ import { useEntityProp } from '@wordpress/core-data';
 import { useState } from '@wordpress/element';
 import { __experimentalGetSettings, dateI18n } from '@wordpress/date';
 import {
-	AlignmentToolbar,
+	AlignmentControl,
 	BlockControls,
 	InspectorControls,
 	useBlockProps,
 } from '@wordpress/block-editor';
 import {
-	ToolbarGroup,
 	ToolbarButton,
 	ToggleControl,
 	Popover,
@@ -91,8 +90,8 @@ export default function PostDateEdit( { attributes, context, setAttributes } ) {
 	}
 	return (
 		<>
-			<BlockControls>
-				<AlignmentToolbar
+			<BlockControls group="block">
+				<AlignmentControl
 					value={ textAlign }
 					onChange={ ( nextAlign ) => {
 						setAttributes( { textAlign: nextAlign } );
@@ -100,17 +99,15 @@ export default function PostDateEdit( { attributes, context, setAttributes } ) {
 				/>
 
 				{ date && (
-					<ToolbarGroup>
-						<ToolbarButton
-							icon={ edit }
-							title={ __( 'Change Date' ) }
-							onClick={ () =>
-								setIsPickerOpen(
-									( _isPickerOpen ) => ! _isPickerOpen
-								)
-							}
-						/>
-					</ToolbarGroup>
+					<ToolbarButton
+						icon={ edit }
+						title={ __( 'Change Date' ) }
+						onClick={ () =>
+							setIsPickerOpen(
+								( _isPickerOpen ) => ! _isPickerOpen
+							)
+						}
+					/>
 				) }
 			</BlockControls>